+++ fastapi_zero/app.py
@@ -1,163 +1,17 @@
 # biblioteca padrão
 from http import HTTPStatus
 
-from fastapi import Depends, FastAPI, HTTPException
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy import select
-from sqlalchemy.orm import Session
+from fastapi import FastAPI
 
-from fastapi_zero.database import get_session
-from fastapi_zero.models import User
-from fastapi_zero.schemas import (
-    Message,
-    UserList,
-    UserPublic,
-    UserSchema,
-)
-from fastapi_zero.security import (
-    create_acess_token,
-    get_current_user,
-    get_password_hash,
-    verify_password,
-)
-
-# from fastapi_zero.utils.validators import (
-#     #check_already_registered_email,
-#     check_user_not_found,
-# )
+from fastapi_zero.routers import auth, users
+from fastapi_zero.schemas import Message
 
 app = FastAPI()
 
-
-@app.post('/users/', status_code=HTTPStatus.CREATED, response_model=UserPublic)
-def create_users(user: UserSchema, session: Session = Depends(get_session)):
-    db_user = session.scalar(
-        select(User).where(
-            (User.username == user.username) | (User.email == user.email)
-        )
-    )
-
-    if db_user:
-        if db_user.username == user.username:
-            raise HTTPException(
-                status_code=HTTPStatus.CONFLICT,
-                detail='Usuário já cadastrado',
-            )
-
-        if db_user.email == user.email:
-            raise HTTPException(
-                status_code=HTTPStatus.CONFLICT,
-                detail='Email já cadastrado',
-            )
-
-    db_user = User(
-        username=user.username,
-        password=get_password_hash(user.password),
-        email=user.email,
-    )
-
-    session.add(db_user)
-    session.commit()
-    session.refresh(db_user)
-
-    return db_user
+app.include_router(users.router)
+app.include_router(auth.router)
 
 
-@app.get('/users/', response_model=UserList)
-def read_users(
-    limit: int = 10, skip: int = 0, session: Session = Depends(get_session)
-):
-    user = session.scalars(select(User).limit(limit).offset(skip)).all()
-    return {'users': user}
-
-
-@app.get('/users/{user_id}', response_model=UserPublic)
-def read_user(user_id: int, session: Session = Depends(get_session)):
-    user_with_id = session.scalar(select(User).where(User.id == user_id))
-
-    if not user_with_id:
-        raise HTTPException(
-            status_code=HTTPStatus.NOT_FOUND, detail='Usuário não encontrado'
-        )
-
-    return user_with_id
-
-
-# PENDENTE: TENTAR COLOCAR ID INVALIDO
-@app.put('/users/{user_id}', response_model=UserPublic)
-def update_user(
-    user_id: int,
-    user: UserSchema,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=HTTPStatus.FORBIDDEN,
-            detail='Você não tem permissão para editar esse usuário',
-        )
-
-    current_user.username = user.username
-    current_user.password = get_password_hash(user.password)
-    current_user.email = user.email
-
-    session.add(current_user)
-    session.commit()
-    session.refresh(current_user)
-
-    return current_user
-
-
-@app.delete('/users/{user_id}', response_model=Message)
-def delete_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=HTTPStatus.FORBIDDEN,
-            detail='Você não tem permissão para deletar esse usuário',
-        )
-
-    session.delete(current_user)
-    session.commit()
-
-    return {'message': 'Usuário deletado com sucesso'}
-
-
-@app.get('/users/{user_id}', response_model=UserPublic)
-def read_user__exercicio(
-    user_id: int, session: Session = Depends(get_session)
-):
-    db_user = session.scalar(select(User).where(User.id == user_id))
-
-    if not db_user:
-        raise HTTPException(
-            status_code=HTTPStatus.NOT_FOUND, detail='Usuário não encontrado'
-        )
-
-    return db_user
-
-
-@app.post('/token')
-def login_for_acess_token(
-    form_data: OAuth2PasswordRequestForm = Depends(),
-    session: Session = Depends(get_session),
-):
-    user = session.scalar(select(User).where(User.email == form_data.username))
-
-    if not user or not verify_password(form_data.password, user.password):
-        raise HTTPException(
-            status_code=HTTPStatus.UNAUTHORIZED,
-            detail='Usuário ou senha inválidos',
-        )
-
-    # a Claim sub é o padrão do JWT para o usuário
-    # o token é gerado com o email do usuário
-    access_token = create_acess_token(data={'sub': user.email})
-
-    return {
-        'access_token': access_token,
-        'token_type': 'Bearer',  # bearer é o padrão do OAuth2
-    }
+@app.get('/', status_code=HTTPStatus.OK, response_model=Message)
+def read_root():
+    return {'Hello': 'World'}
