+++ fastapi_zero/routers/auth.py
@@ -1,0 +1,39 @@
+from http import HTTPStatus
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from fastapi_zero.database import get_session
+from fastapi_zero.models import User
+from fastapi_zero.security import (
+    create_acess_token,
+    verify_password,
+)
+
+router = APIRouter(prefix='/auth', tags=['auth'])
+
+T_Session = Annotated[Session, Depends(get_session)]
+T_OAuthForm = Annotated[OAuth2PasswordRequestForm, Depends()]
+
+
+@router.post('/token')
+def login_for_acess_token(session: T_Session, form_data: T_OAuthForm):
+    user = session.scalar(select(User).where(User.email == form_data.username))
+
+    if not user or not verify_password(form_data.password, user.password):
+        raise HTTPException(
+            status_code=HTTPStatus.UNAUTHORIZED,
+            detail='Usuário ou senha inválidos',
+        )
+
+    # a Claim sub é o padrão do JWT para o usuário
+    # o token é gerado com o email do usuário
+    access_token = create_acess_token(data={'sub': user.email})
+
+    return {
+        'access_token': access_token,
+        'token_type': 'Bearer',  # bearer é o padrão do OAuth2
+    }
