+++ pyproject.toml
@@ -47,7 +47,7 @@ pre_format = 'ruff check --fix'
 format = 'ruff format'
 run = 'fastapi dev fastapi_zero/app.py'
 pre_test = 'task lint'
-test = 'pytest -s -x --cov=fastapi_zero -vv'
+test = 'pytest -s --cov=fastapi_zero -vv'
 post_test = 'coverage html'
 
 [build-system]
