+++ tests/test_auth.py
@@ -1,0 +1,17 @@
+from http import HTTPStatus
+
+
+def test_get_token(client, user):
+    response = client.post(
+        'auth/token',
+        data={
+            'username': user.email,
+            'password': user.clean_password,
+        },
+    )
+
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert token['token_type'] == 'Bearer'
+    assert 'access_token' in token
